pyproject.toml
[project]
name = "language-engineer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.13.0",
    "bert-score>=0.3.13",
    "bitsandbytes>=0.49.2",
    "datasets>=4.8.2",
    "peft>=0.18.1",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.3",
    "rouge-score>=0.1.2",
    "sacrebleu>=2.6.0",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
    "trl>=0.29.0",
    "yake>=0.4.8",
    "sumy>=0.11.0",
    "nltk>=3.9.0",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cu130" }]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "pre-commit>=4.5.1",
]
